+++ src/nodes/features.h
@@ -94,7 +94,7 @@ private:
   }
 
   bool setMode(dc1394feature_info_t *finfo, dc1394feature_mode_t mode);
-  void setOff(dc1394feature_info_t *finfo);
+  void setPower(dc1394feature_info_t *finfo, dc1394switch_t on_off);
   void updateIfChanged(dc1394feature_t feature,
                        int old_control, int *control,
                        double old_value, double *value);